# tests/contract/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_document_v1_documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "PDF, CSV, Markdown, text, or log file",
            "title": "File",
            "type": "string"
          },
          "force_reindex": {
            "default": false,
            "title": "Force Reindex",
            "type": "boolean"
          },
          "metadata_json": {
            "default": "{}",
            "title": "Metadata Json",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_v1_documents_post",
        "type": "object"
      },
      "DependencyHealth": {
        "additionalProperties": false,
        "description": "Safe readiness state for one required dependency.",
        "properties": {
          "latency_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "status": {
            "$ref": "#/components/schemas/HealthStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "DependencyHealth",
        "type": "object"
      },
      "DocumentDeletionResponse": {
        "additionalProperties": false,
        "description": "Idempotent deletion status and honest backup-retention boundary.",
        "properties": {
          "document_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Document Id",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Request Id",
            "type": "string"
          },
          "retention_notice": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Retention Notice",
            "type": "string"
          },
          "schema_version": {
            "default": "v1",
            "pattern": "^v1$",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          }
        },
        "required": [
          "document_id",
          "status",
          "retention_notice",
          "request_id"
        ],
        "title": "DocumentDeletionResponse",
        "type": "object"
      },
      "DocumentListResponse": {
        "additionalProperties": false,
        "description": "Cursor-paginated document listing.",
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentMetadata"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "schema_version": {
            "default": "v1",
            "pattern": "^v1$",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "items",
          "has_more"
        ],
        "title": "DocumentListResponse",
        "type": "object"
      },
      "DocumentMetadata": {
        "additionalProperties": false,
        "description": "Tenant-safe document metadata.",
        "properties": {
          "checksum_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Checksum Sha256",
            "type": "string"
          },
          "content_type": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Document Id",
            "type": "string"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "maxProperties": 64,
            "title": "User Metadata",
            "type": "object"
          }
        },
        "required": [
          "document_id",
          "filename",
          "content_type",
          "checksum_sha256",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentMetadata",
        "type": "object"
      },
      "DocumentStatus": {
        "description": "Externally visible ingestion and deletion states.",
        "enum": [
          "accepted",
          "parsing",
          "chunking",
          "indexing",
          "ready",
          "failed",
          "deleting",
          "deleted"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentStatusResponse": {
        "additionalProperties": false,
        "description": "Current processing state without unsafe parser or provider details.",
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentMetadata"
          },
          "failure_code": {
            "anyOf": [
              {
                "maxLength": 64,
                "pattern": "^[a-z0-9_]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "schema_version": {
            "default": "v1",
            "pattern": "^v1$",
            "title": "Schema Version",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionManifest"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/DocumentWarning"
            },
            "maxItems": 100,
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "document",
          "versions"
        ],
        "title": "DocumentStatusResponse",
        "type": "object"
      },
      "DocumentUploadResponse": {
        "additionalProperties": false,
        "description": "Acknowledgement returned after a durable ingestion job is accepted.",
        "properties": {
          "document_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Document Id",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "default": "v1",
            "pattern": "^v1$",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus",
            "default": "accepted"
          }
        },
        "required": [
          "document_id",
          "request_id"
        ],
        "title": "DocumentUploadResponse",
        "type": "object"
      },
      "DocumentWarning": {
        "additionalProperties": false,
        "description": "Safe, non-terminal processing warning.",
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[a-z0-9_]+$",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "DocumentWarning",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "description": "Stable public error body.",
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[a-z0-9_]+$",
            "title": "Code",
            "type": "string"
          },
          "field_violations": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "maxItems": 32,
            "title": "Field Violations",
            "type": "array"
          },
          "message": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Versioned error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "schema_version": {
            "default": "v1",
            "pattern": "^v1$",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FieldViolation": {
        "additionalProperties": false,
        "description": "Bounded validation detail safe to return to a caller.",
        "properties": {
          "field": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Field",
            "type": "string"
          },
          "message": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldViolation",
        "type": "object"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "description": "Liveness or readiness response.",
        "properties": {
          "dependencies": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DependencyHealth"
            },
            "maxProperties": 16,
            "title": "Dependencies",
            "type": "object"
          },
          "revision": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Revision",
            "type": "string"
          },
          "schema_version": {
            "default": "v1",
            "pattern": "^v1$",
            "title": "Schema Version",
            "type": "string"
          },
          "service": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Service",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/HealthStatus"
          }
        },
        "required": [
          "status",
          "service",
          "revision"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "HealthStatus": {
        "description": "Service health states.",
        "enum": [
          "ok",
          "not_ready"
        ],
        "title": "HealthStatus",
        "type": "string"
      },
      "VersionManifest": {
        "additionalProperties": false,
        "description": "Versions needed to reproduce a query or ingestion result.",
        "properties": {
          "app_revision": {
            "maxLength": 128,
            "minLength": 1,
            "title": "App Revision",
            "type": "string"
          },
          "chunker_version": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunker Version"
          },
          "dense_model_revision": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Model Revision"
          },
          "generator_model_revision": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generator Model Revision"
          },
          "index_schema_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Index Schema Version",
            "type": "string"
          },
          "parser_version": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "prompt_version": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "reranker_model_revision": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reranker Model Revision"
          },
          "sparse_model_revision": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sparse Model Revision"
          }
        },
        "required": [
          "app_revision",
          "index_schema_version"
        ],
        "title": "VersionManifest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Tenant-safe hybrid document retrieval and grounded answers.",
    "title": "Hybrid RAG Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health/live": {
      "get": {
        "description": "Report only process and event-loop liveness.",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Live",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "description": "Report configuration and Qdrant connectivity without model calls.",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ready",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/documents": {
      "get": {
        "operationId": "list_documents_v1_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "upload_document_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_v1_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDeletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Document",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "get_document_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "documents"
        ]
      }
    }
  }
}